# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Relaycode API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "operationId": "events_stream",
        "responses": {
          "200": {
            "description": "Server-Sent Events stream"
          }
        }
      }
    },
    "/api/prompts": {
      "get": {
        "tags": [
          "Prompts"
        ],
        "operationId": "list_prompts",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Prompt"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "operationId": "list_transactions",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/bulk": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "operationId": "bulk_update_transactions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/{id}/files/reapply": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "operationId": "reapply_single_file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Transaction ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReapplyFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "File reapply initiated"
          }
        }
      }
    },
    "/api/transactions/{id}/reapply-failed": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "operationId": "reapply_all_failed_files",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Transaction ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reapply all failed files initiated"
          }
        }
      }
    },
    "/api/transactions/{id}/status": {
      "patch": {
        "tags": [
          "Transactions"
        ],
        "operationId": "update_transaction_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Transaction ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BulkActionRequest": {
        "type": "object",
        "required": [
          "ids",
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/TransactionStatus"
          },
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "BulkActionResponse": {
        "type": "object",
        "required": [
          "success",
          "updatedIds"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "updatedIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "FileApplyStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPLYING",
          "APPLIED",
          "FAILED"
        ]
      },
      "FileStatus": {
        "type": "string",
        "enum": [
          "modified",
          "created",
          "deleted",
          "renamed"
        ]
      },
      "FileStatusEvent": {
        "type": "object",
        "required": [
          "transactionId",
          "filePath",
          "applyStatus"
        ],
        "properties": {
          "applyStatus": {
            "$ref": "#/components/schemas/FileApplyStatus"
          },
          "errorMessage": {
            "type": [
              "string",
              "null"
            ]
          },
          "filePath": {
            "type": "string"
          },
          "transactionId": {
            "type": "string"
          }
        }
      },
      "Prompt": {
        "type": "object",
        "required": [
          "id",
          "title",
          "content",
          "timestamp",
          "status"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PromptStatus"
          },
          "timestamp": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "PromptStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "ACTIVE",
          "COMPLETED",
          "ARCHIVED"
        ]
      },
      "ReapplyFileRequest": {
        "type": "object",
        "required": [
          "filePath"
        ],
        "properties": {
          "filePath": {
            "type": "string"
          }
        }
      },
      "SimulationEvent": {
        "type": "object",
        "required": [
          "transactionId",
          "status",
          "timestamp"
        ],
        "properties": {
          "progress": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/TransactionStatus"
          },
          "timestamp": {
            "type": "string"
          },
          "transactionId": {
            "type": "string"
          }
        }
      },
      "SimulationScenario": {
        "type": "string",
        "enum": [
          "fast-success",
          "simulated-failure",
          "long-running",
          "partial-failure"
        ]
      },
      "Transaction": {
        "type": "object",
        "required": [
          "id",
          "status",
          "description",
          "timestamp",
          "createdAt",
          "promptId",
          "author",
          "blocks",
          "files",
          "provider",
          "model",
          "cost",
          "tokens",
          "reasoning"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionBlock"
            }
          },
          "cost": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionFile"
            }
          },
          "id": {
            "type": "string"
          },
          "isChainRoot": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "model": {
            "type": "string"
          },
          "parentId": {
            "type": [
              "string",
              "null"
            ]
          },
          "promptId": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "reasoning": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TransactionStatus"
          },
          "timestamp": {
            "type": "string"
          },
          "tokens": {
            "type": "string"
          }
        }
      },
      "TransactionBlock": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "file": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TransactionFile"
              }
            ]
          },
          "type": {
            "type": "string"
          }
        }
      },
      "TransactionFile": {
        "type": "object",
        "required": [
          "path",
          "status",
          "language",
          "diff"
        ],
        "properties": {
          "applyStatus": {
            "$ref": "#/components/schemas/FileApplyStatus"
          },
          "diff": {
            "type": "string"
          },
          "errorMessage": {
            "type": [
              "string",
              "null"
            ]
          },
          "language": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/FileStatus"
          }
        }
      },
      "TransactionStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPLYING",
          "APPLIED",
          "PARTIALLYAPPLIED",
          "COMMITTED",
          "REVERTED",
          "FAILED"
        ]
      },
      "UpdateStatusRequest": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "scenario": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SimulationScenario"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/TransactionStatus"
          }
        }
      }
    }
  }
}